backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
supabase>=2.9.0
python-multipart>=0.0.12
pdfplumber>=0.11.0
python-docx>=1.1.0
reportlab>=4.2.0
python-jose[cryptography]>=3.3.0
openpyxl>=3.1.0
python-pptx>=1.0.0
